pyproject.toml
[build-system]
requires = ["poetry>=1.2.0b2"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "harvest_kit_hiwi"
version = "0.1.0"
description = "Automatically export harvest time tracking to kit hiwi template"
license = "MIT"
authors = ["Jonas Teufel <jonseb1998@gmail.com>"]
maintainers = ["Jonas Teufel <jonseb1998@gmail.com>"]
readme = "README.rst"
keywords = ["automation", "KIT", "template", "time tracking"]
packages = [
    { include = "harvest_kit_hiwi" },
    { include = "tests" },
]
include = [
    "harvest_kit_hiwi/VERSION",
    "harvest_kit_hiwi/config.yml",
]

[tool.poetry.scripts]
harvest_hiwi = "harvest_kit_hiwi.cli:main"

[tool.poetry.dependencies]
python = ">=3.8.0"
click = ">=7.1.2"
svgutils = ">=0.3.4"
cairosvg = ">=2.5.2"
requests = ">=2.28.1"
python-dateutil = ">=2.8.2"
pyyaml = ">=6.0"
python-decouple = ">=3.6"

[tool.poetry.dev-dependencies]
sphinx = "5.0.2"
pytest = "7.1.3"
poetry_bumpversion = ">=0.2.0"

[tool.poetry_bumpversion.file."harvest_kit_hiwi/VERSION"]
search = "{current_version}"
replace = "{new_version}"

[tool.poetry_bumpversion.file."README.rst"]
search = "{current_version}"
replace = "{new_version}"
